+++ CHANGELOG.md
@@ -42,6 +42,7 @@
 - (Feature) Deprecate AF Mode
 - (Maintenance) Switch License to 2025
 - (Feature) Migrate Storage V1 to CE
+- (Feature) Improve Helm Chart Manager
 
 ## [1.2.43](https://github.com/arangodb/kube-arangodb/tree/1.2.43) (2024-10-14)
 - (Feature) ArangoRoute CRD